requirements.txt
# Ingestion
hypersync
psycopg2-binary
sqlalchemy
python-dotenv
pandas
pyarrow

# dbt
dbt-core
dbt-postgres

# Analysis
lifelines
matplotlib

# Dashboard
dash
plotly

# Agents (deferred)
# openai-agents
